Complex types - functions

+++ 0x00-python_variable_annotations/8-make_multiplier.py
@@ -12,7 +12,7 @@ def mult(a: float) -> float:
         Return:
             multiple of a and b
     """
-    return a * a
+    return float(a * a)
 
 
 def make_multiplier(multiplier: float) -> Callable[[float], float]:
